Renamed folder and adjusted namespaces
