+++ src/Sample/Configuration/QuestionConfigurationSetting.cs
@@ -1,6 +1,6 @@
 ﻿using ConfigInjector;
 
-namespace Sample.ConfigSettings
+namespace Sample.Configuration
 {
     public class QuestionConfigurationSetting: ConfigurationSetting<string>
     {
